CoinScan/requirements.txt
numpy>=1.23.0,<2.0
opencv-python>=4.5.5.62,<5.0
Pillow>=9.0.0,<10.0